refact: structured firebase functions
