+++ events/guildMemberUpdate.js
@@ -1,6 +1,5 @@
-// eslint-disable-next-line no-unused-vars
-const Discord = require('discord.js');
-const { updateUser } = require('../firebase');
+const Discord = require('discord.js'); // eslint-disable-line no-unused-vars
+const { updateUser } = require('../firebase/member');
 
 module.exports = {
     name: 'guildMemberUpdate',
